backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
python-multipart==0.0.20
pydantic-settings==2.7.1
pydantic==2.10.6



 